Simplify Arduino SHT interface, add auto-detect
As part of the simplification, all sensors were unified in one source
and one header file. The interface remains largely the same but
instead of providing one class for each sensor, there is now a single
SHTSensor class to access all sensors.

The individual classes are still used internally.

By default (when initialized without parameters), SHTSensor will probe
the bus for available sensors of type SHT3x, SHT3xAlt (Address 0x45)
and the SHTC1 family (including SHTW1 and SHTW2)

The SHT3xAnalog must be instantiated separately since it requires the
ADC pin parameters, but for convenience, it can be passed to the
SHTSensor class to provide a unified interface for all sensors.

The examples and README.md are updated.

The code was tested with an SHT3x and an SHTC1. The SHT3xAnalog code
paths were tested without an attached sensor.